Fix cross-case method contamination when running multiple cases in the same REPL
When a user runs case A then case B in the same Julia session, methods defined
in A's user_source.jl/user_primitives.jl persist in the global method table
and get called instead of B's methods, causing incorrect results silently.

Two-part fix:

1. src/run.jl: Before reloading case files, explicitly delete all methods of
   user_flux!, user_source!, user_bc_dirichlet!, user_primitives!, user_uout!.
   These functions have no definitions in src/ — every method comes from a case
   file — so a full delete-then-reload is safe and sufficient.

2. 108 case files (user_source.jl, user_primitives.jl, user_bc.jl across all
   problem directories): Replace runtime ET == TOTAL() / ET == PERT() checks
   with proper ::TOTAL / ::PERT type dispatch, and remove ::TInt / ::Int64 type
   annotations on the npoin argument. This makes methods unambiguous when
   reloaded into a clean method table, and ensures Julia dispatches correctly
   without relying on stale method ordering.

+++ problems/CompEuler/LESICP2/user_source.jl
@@ -1,7 +1,7 @@
 function user_source!(S,
                       q, 
                       qe,
-                      npoin::Int64,
+                      npoin,
                       ::CL, ::TOTAL;
                       neqs=1,
                       x=0.0,
