+++ problems/CompEuler/nozzleanderson/user_source.jl
@@ -1,7 +1,7 @@
 function user_source!(S,
                       q, 
                       qe,
-                      npoin::Int64,
+                      npoin,
                       ::CL, ::TOTAL;
                       neqs=1, x=1, y=1, xmax=1, xmin=1, ymax=1)
