+++ problems/Elliptic/elementLearning/user_source.jl
@@ -1,7 +1,7 @@
 function user_source!(S,
                       q, 
                       qe,
-                      npoin::Int64,
+                      npoin,
                       ::CL, ::TOTAL;
                       neqs=1,
                       x=1.0, y=1.0,
